+++ packages/terra-core-docs/CHANGELOG.md
@@ -9,6 +9,7 @@
   * Migrate doc from terra-card.
   * Migrate doc from terra-button.
   * Migrate doc from terra-button-group.
+  * Migrate docs from terra-form-field.
 
 ## 1.11.0 - (October 8, 2021)